refactor(resources): eliminate ResourceRegistry interface
- Delete unnecessary interface abstraction
- Rename ResourceProviderRegistry to ResourceRegistry
- Update imports in MCPDependencies, MCPIntegrationService, McpMethodHandlers
- Simplify DI configuration
- Follow same pattern as Phase 2 ToolRegistry elimination
